requirements.txt
dash>=2.0.0
numpy>=1.21.2
pandas>=1.3.3
umap-learn
scikit-learn
